refactor: CartAdapter ListAdapter 활용해서 구현하도록 수정
