+++ app/src/main/java/woowacourse/shopping/presentation/cart/CartDiffUtil.kt
@@ -1,0 +1,16 @@
+package woowacourse.shopping.presentation.cart
+
+import androidx.recyclerview.widget.DiffUtil
+import woowacourse.shopping.presentation.CartItemUiModel
+
+object CartDiffUtil : DiffUtil.ItemCallback<CartItemUiModel>() {
+    override fun areItemsTheSame(
+        oldItem: CartItemUiModel,
+        newItem: CartItemUiModel,
+    ): Boolean = oldItem.id == newItem.id
+
+    override fun areContentsTheSame(
+        oldItem: CartItemUiModel,
+        newItem: CartItemUiModel,
+    ): Boolean = oldItem == newItem
+}
